+++ Sources/PopupBridge/POPPopupBridge.swift
@@ -1,64 +1,48 @@
 import Foundation
-import SafariServices
 import WebKit
+import AuthenticationServices
 
 @objcMembers
-public class POPPopupBridge: NSObject, WKScriptMessageHandler, SFSafariViewControllerDelegate {
+public class POPPopupBridge: NSObject, WKScriptMessageHandler {
+
+    /// Exposed for testing
+    var returnedWithURL: Bool = false
     
     // MARK: - Private Properties
     
     private let messageHandlerName = "POPPopupBridge"
-    private let hostName = "popupbridgev1"
-    
+    private let hostName = "popupbridgev1"    
     private let webView: WKWebView
-    private let urlScheme: String
-    private let delegate: POPPopupBridgeDelegate
-    private var safariViewController: SFSafariViewController?
+    private var webAuthenticationSession: WebAuthenticationSession = WebAuthenticationSession()
     
-    private static var returnBlock: ((URL) -> Bool)?
+    private var returnBlock: ((URL) -> Void)? = nil
     
-    // MARK: - Initializer
+    // MARK: - Initializers
         
     /// Initialize a Popup Bridge.
     /// - Parameters:
     ///   - webView: The web view to add a script message handler to. Do not change the web view's configuration or user content controller after initializing Popup Bridge.
-    ///   - urlScheme: The URL Scheme that you have registered in your Info.plist.
-    ///   - delegate: A delegate that presents and dismisses the pop-up (a SFSafariViewController).
-    public init(
-        webView: WKWebView,
-        urlScheme: String,
-        delegate: POPPopupBridgeDelegate
-    ) {
+    public init(webView: WKWebView) {
         self.webView = webView
-        self.urlScheme = urlScheme
-        self.delegate = delegate
         
         super.init()
-        
-        configureWebView(scheme: urlScheme)
-        
-        POPPopupBridge.returnBlock = { (url : URL) -> Bool in
+
+        configureWebView()
+                
+        returnBlock = { url in
             guard let script = self.constructJavaScriptCompletionResult(returnURL: url) else {
-                return false
+                return
             }
-            
+
             self.injectWebView(webView: webView, withJavaScript: script)
-            return true
+            return
         }
     }
-    
-    // MARK: - Public Methods
-    
-    /// Handle completion of the popup flow by calling this method from either
-    /// your scene:openURLContexts: scene delegate method or
-    /// your application:openURL:sourceApplication:annotation: app delegate method.
-    @objc(openURL:)
-    public static func open(url: URL) -> Bool {
-        if let returnBlock {
-            return returnBlock(url)
-        } else {
-            return false
-        }
+
+    /// Exposed for testing
+    convenience init(webView: WKWebView, webAuthenticationSession: WebAuthenticationSession) {
+        self.init(webView: webView)
+        self.webAuthenticationSession = webAuthenticationSession
     }
     
     // MARK: - Internal Methods
@@ -66,17 +50,15 @@ public class POPPopupBridge: NSObject, WKScriptMessageHandler, SFSafariViewContr
     /// Exposed for testing
     /// 
     /// Constructs custom JavaScript to be injected into the merchant's WKWebView, based on redirectURL details from the SFSafariViewController pop-up result.
-    /// - Parameter url: returnURL from the result of the SFSafariViewController being dismissed. Provided via the merchant's AppDelegate or SceneDelegate.
+    /// - Parameter url: returnURL from the result of the ASWebAuthenticationSession.
     /// - Returns: JavaScript formatted completion.
     func constructJavaScriptCompletionResult(returnURL: URL) -> String? {
         guard let urlComponents = URLComponents(url: returnURL, resolvingAgainstBaseURL: false),
-              urlComponents.scheme?.caseInsensitiveCompare(urlScheme) == .orderedSame,
+              urlComponents.scheme?.caseInsensitiveCompare(PopupBridgeConstants.callbackURLScheme) == .orderedSame,
               urlComponents.host?.caseInsensitiveCompare(hostName) == .orderedSame
         else {
             return nil
         }
-
-        self.dismissSafariViewController()
 
         let queryItems = urlComponents.queryItems?.reduce(into: [:]) { partialResult, queryItem in
             partialResult[queryItem.name] = queryItem.value
@@ -100,11 +82,11 @@ public class POPPopupBridge: NSObject, WKScriptMessageHandler, SFSafariViewContr
     
     /// Injects custom JavaScript into the merchant's webpage.
     /// - Parameter scheme: the url scheme provided by the merchant
-    private func configureWebView(scheme: String) {
+    private func configureWebView() {
         webView.configuration.userContentController.add(self, name: messageHandlerName)
         
         let javascript = PopupBridgeUserScript(
-            scheme: scheme,
+            scheme: PopupBridgeConstants.callbackURLScheme,
             scriptMessageHandlerName: messageHandlerName,
             host: hostName
         ).rawJavascript
@@ -117,35 +99,13 @@ public class POPPopupBridge: NSObject, WKScriptMessageHandler, SFSafariViewContr
         
         webView.configuration.userContentController.addUserScript(script)
     }
-    
-    private func dismissSafariViewController() {
-        if let safariViewController {
-            delegate.popupBridge(self, requestsDismissalOfViewController: safariViewController)
-        }
-    }
-    
+
     private func injectWebView(webView: WKWebView, withJavaScript script: String) {
         webView.evaluateJavaScript(script) { _, error in
             if let error {
                 NSLog("Error: PopupBridge requires onComplete callback. Details: %@", error.localizedDescription)
             }
         }
-    }
-        
-    // MARK: - SFSafariViewControllerDelegate conformance
-    
-    /// :nodoc: This method is not covered by Semantic Versioning. Do not use.
-    ///
-    /// Called when the user exits the pop-up (SFSafariViewController) by clicking "Done"
-    public func safariViewControllerDidFinish(_ controller: SFSafariViewController) {
-        let script = """
-            if (typeof window.popupBridge.onCancel === 'function') {\
-              window.popupBridge.onCancel();\
-            } else {\
-              window.popupBridge.onComplete(null, null);\
-            }
-            """
-        injectWebView(webView: webView, withJavaScript: script)
     }
     
     // MARK: - WKScriptMessageHandler conformance
@@ -160,22 +120,42 @@ public class POPPopupBridge: NSObject, WKScriptMessageHandler, SFSafariViewContr
                 return
             }
             
-            if let urlString = script.url,
-               let url = URL(string: urlString) {
-                dismissSafariViewController()
-                
-                delegate.popupBridge?(self, willOpenURL: url)
-                
-                let viewController = SFSafariViewController(url: url)
-                safariViewController = viewController
-                safariViewController?.delegate = self
-                
-                self.delegate.popupBridge(self, requestsPresentationOfViewController: viewController)
-                return
-            } else if let name = script.message?.name {
-                delegate.popupBridge?(self, receivedMessage: name, data: script.message?.data)
-                return
+            if let urlString = script.url, let url = URL(string: urlString) {
+                webAuthenticationSession.start(url: url, context: self) { url, _ in
+                    if let url, let returnBlock = self.returnBlock {
+                        self.returnedWithURL = true
+                        returnBlock(url)
+                        return
+                    }
+                } sessionDidCancel: { [self] in
+                    let script = """
+                        if (typeof window.popupBridge.onCancel === 'function') {\
+                            window.popupBridge.onCancel();\
+                        } else {\
+                            window.popupBridge.onComplete(null, null);\
+                        }
+                    """
+
+                    injectWebView(webView: webView, withJavaScript: script)
+                    return
+                }
             }
+        }
+    }
+}
+
+// MARK: - ASWebAuthenticationPresentationContextProviding conformance
+
+extension POPPopupBridge: ASWebAuthenticationPresentationContextProviding {
+
+    public func presentationAnchor(for session: ASWebAuthenticationSession) -> ASPresentationAnchor {
+        if #available(iOS 15, *) {
+            let firstScene = UIApplication.shared.connectedScenes.first as? UIWindowScene
+            let window = firstScene?.windows.first { $0.isKeyWindow }
+            return window ?? ASPresentationAnchor()
+        } else {
+            let window = UIApplication.shared.windows.first { $0.isKeyWindow }
+            return window ?? ASPresentationAnchor()
         }
     }
 }