+++ UnitTests/PopupBridge_UnitTests.swift
@@ -1,22 +1,21 @@
 import XCTest
 import WebKit
-import SafariServices
 @testable import PopupBridge
 
 final class PopupBridge_UnitTests: XCTestCase, WKNavigationDelegate {
 
     let scriptMessageHandlerName: String = "POPPopupBridge"
     let returnURL: String = "com.braintreepayments.popupbridgeexample"
+    let mockWebAuthenticationSession = MockWebAuthenticationSession()
 
     var webViewReadyBlock: (Void)?
 
     func testInit_addsUserScript() {
         let webView = WKWebView()
-        let delegate = MockPopupBridgeDelegate()
 
         XCTAssertEqual(webView.configuration.userContentController.userScripts.count, 0)
 
-        let _ = POPPopupBridge(webView: webView, urlScheme: returnURL, delegate: delegate)
+        let _ = POPPopupBridge(webView: webView)
 
         XCTAssertEqual(webView.configuration.userContentController.userScripts.count, 1)
 
@@ -26,21 +25,21 @@ final class PopupBridge_UnitTests: XCTestCase, WKNavigationDelegate {
         XCTAssertTrue(userScript.isForMainFrameOnly)
     }
 
+
     func testInit_addsScriptMessageHandler() {
         let configuration = WKWebViewConfiguration()
         let mockUserContentController = MockUserContentController()
-        let delegate = MockPopupBridgeDelegate()
 
         configuration.userContentController = mockUserContentController
 
         let webView = WKWebView(frame: CGRect(), configuration: configuration)
-        let pub = POPPopupBridge(webView: webView, urlScheme: returnURL, delegate: delegate)
+        let pub = POPPopupBridge(webView: webView)
 
         XCTAssertEqual(mockUserContentController.scriptMessageHandler as? POPPopupBridge, pub)
         XCTAssertEqual(mockUserContentController.name, scriptMessageHandlerName)
     }
 
-    func testReceiveScriptMessage_whenMessageContainsURL_requestsPresentationOfSafariViewController() {
+    func testReceiveScriptMessage_whenMessageContainsURL_startsWebAuthenticationSession() {
         let stubMessageBody: [String: String] = ["url": "http://example.com/?hello=world"]
         let stubMessageName = scriptMessageHandlerName
         let stubMessage = MockScriptMessage()
@@ -49,17 +48,16 @@ final class PopupBridge_UnitTests: XCTestCase, WKNavigationDelegate {
 
         let configuration = WKWebViewConfiguration()
         let mockUserContentController = MockUserContentController()
-        let delegate = MockPopupBridgeDelegate()
 
         configuration.userContentController = mockUserContentController
 
         let webView = WKWebView(frame: CGRect(), configuration: configuration)
-        let pub = POPPopupBridge(webView: webView, urlScheme: returnURL, delegate: delegate)
+        let pub = POPPopupBridge(webView: webView, webAuthenticationSession: mockWebAuthenticationSession)
 
+        mockWebAuthenticationSession.cannedResponseURL = URL(string: "http://example.com/?hello=world")
         pub.userContentController(WKUserContentController(), didReceive: stubMessage)
-        delegate.popupBridge(pub, requestsDismissalOfViewController: UIViewController())
 
-        XCTAssertTrue(delegate.didRequestPresentationOfViewController)
+        XCTAssertTrue(pub.returnedWithURL)
     }
 
     func testReceiveScriptMessage_whenMessageContainsURL_informsDelegateThatURLWillBeLoaded() {
@@ -71,12 +69,11 @@ final class PopupBridge_UnitTests: XCTestCase, WKNavigationDelegate {
 
         let configuration = WKWebViewConfiguration()
         let mockUserContentController = MockUserContentController()
-        let delegate = MockPopupBridgeDelegate()
 
         configuration.userContentController = mockUserContentController
 
         let webView = WKWebView(frame: CGRect(), configuration: configuration)
-        let pub = POPPopupBridge(webView: webView, urlScheme: returnURL, delegate: delegate)
+        let pub = POPPopupBridge(webView: webView)
 
         pub.userContentController(WKUserContentController(), didReceive: stubMessage)
     }
@@ -90,12 +87,11 @@ final class PopupBridge_UnitTests: XCTestCase, WKNavigationDelegate {
 
         let configuration = WKWebViewConfiguration()
         let mockUserContentController = MockUserContentController()
-        let delegate = MockPopupBridgeDelegate()
 
         configuration.userContentController = mockUserContentController
 
         let webView = WKWebView(frame: CGRect(), configuration: configuration)
-        let pub = POPPopupBridge(webView: webView, urlScheme: returnURL, delegate: delegate)
+        let pub = POPPopupBridge(webView: webView)
 
         pub.userContentController(WKUserContentController(), didReceive: stubMessage)
     }
@@ -109,16 +105,13 @@ final class PopupBridge_UnitTests: XCTestCase, WKNavigationDelegate {
 
         let configuration = WKWebViewConfiguration()
         let mockUserContentController = MockUserContentController()
-        let delegate = MockPopupBridgeDelegate()
 
         configuration.userContentController = mockUserContentController
 
         let webView = WKWebView(frame: CGRect(), configuration: configuration)
-        let pub = POPPopupBridge(webView: webView, urlScheme: returnURL, delegate: delegate)
-        let stubSafari = SFSafariViewController(url: URL(string: "http://example.com")!)
+        let pub = POPPopupBridge(webView: webView)
 
         pub.userContentController(WKUserContentController(), didReceive: stubMessage)
-        pub.safariViewControllerDidFinish(stubSafari)
 
         webView.evaluateJavaScript("""
         "if (typeof window.popupBridge.onCancel === 'function') {"
@@ -129,7 +122,7 @@ final class PopupBridge_UnitTests: XCTestCase, WKNavigationDelegate {
         """)
     }
 
-    func testOpenURL_whenReturnURLHasQueryParams_passesPayloadWithQueryItemsToWebView() {
+    func testConstructJavaScriptCompletionResult_whenReturnURLHasQueryParams_passesPayloadWithQueryItemsToWebView() {
         let stubMessageBody: [String: String] = ["url": "http://example.com/?hello=world"]
         let stubMessageName = scriptMessageHandlerName
         let stubMessage = MockScriptMessage()
@@ -138,21 +131,21 @@ final class PopupBridge_UnitTests: XCTestCase, WKNavigationDelegate {
 
         let configuration = WKWebViewConfiguration()
         let mockUserContentController = MockUserContentController()
-        let delegate = MockPopupBridgeDelegate()
 
         configuration.userContentController = mockUserContentController
 
         let webView = WKWebView(frame: CGRect(), configuration: configuration)
-        let pub = POPPopupBridge(webView: webView, urlScheme: returnURL, delegate: delegate)
-        let mockURL = URL(string: "com.braintreepayments.popupbridgeexample://popupbridgev1/return?something=foo&other=bar")!
-        let _ = POPPopupBridge.open(url: mockURL)
+        let pub = POPPopupBridge(webView: webView, webAuthenticationSession: mockWebAuthenticationSession)
+        let mockURL = URL(string: "sdk.ios.popup-bridge://popupbridgev1/return?something=foo&other=bar")!
+        mockWebAuthenticationSession.cannedResponseURL = mockURL
+
         let expectedResult = "window.popupBridge.onComplete(null, {\"path\":\"\\/return\",\"queryItems\":{\"other\":\"bar\",\"something\":\"foo\"}});"
         let result = pub.constructJavaScriptCompletionResult(returnURL: mockURL)
 
         XCTAssertEqual(result, expectedResult)
     }
 
-    func testOpenURL_whenReturnURLHasNoQueryParams_passesPayloadWithNoQueryItemsToWebView() {
+    func testConstructJavaScriptCompletionResult_whenReturnURLHasNoQueryParams_passesPayloadWithNoQueryItemsToWebView() {
         let stubMessageBody: [String: String] = ["url": "http://example.com/?hello=world"]
         let stubMessageName = scriptMessageHandlerName
         let stubMessage = MockScriptMessage()
@@ -161,112 +154,86 @@ final class PopupBridge_UnitTests: XCTestCase, WKNavigationDelegate {
 
         let configuration = WKWebViewConfiguration()
         let mockUserContentController = MockUserContentController()
-        let delegate = MockPopupBridgeDelegate()
 
         configuration.userContentController = mockUserContentController
 
         let webView = WKWebView(frame: CGRect(), configuration: configuration)
-        let pub = POPPopupBridge(webView: webView, urlScheme: returnURL, delegate: delegate)
-        let mockURL = URL(string: "com.braintreepayments.popupbridgeexample://popupbridgev1/return")!
-        let _ = POPPopupBridge.open(url: mockURL)
+        let pub = POPPopupBridge(webView: webView, webAuthenticationSession: mockWebAuthenticationSession)
+        let mockURL = URL(string: "sdk.ios.popup-bridge://popupbridgev1/return")!
+        mockWebAuthenticationSession.cannedResponseURL = mockURL
+
         let expectedResult = "window.popupBridge.onComplete(null, {\"path\":\"\\/return\",\"queryItems\":{}});"
         let result = pub.constructJavaScriptCompletionResult(returnURL: mockURL)
 
         XCTAssertEqual(result, expectedResult)
     }
 
-    func testOpenURL_whenReturnURLHasURLFragment_passesPayloadWithHashToWebView() {
+    func testConstructJavaScriptCompletionResult_whenReturnURLHasURLFragment_passesPayloadWithHashToWebView() {
         let stubMessageBody: [String: String] = ["url": "http://example.com/?hello=world"]
         let stubMessageName = scriptMessageHandlerName
         let stubMessage = MockScriptMessage()
         stubMessage.body = stubMessageBody
         stubMessage.name = stubMessageName
 
-        let delegate = MockPopupBridgeDelegate()
-        let pub = POPPopupBridge(webView: WKWebView(), urlScheme: returnURL, delegate: delegate)
+        let pub = POPPopupBridge(webView: WKWebView(), webAuthenticationSession: mockWebAuthenticationSession)
 
         pub.userContentController(WKUserContentController(), didReceive: stubMessage)
 
-        let mockURL = URL(string: "com.braintreepayments.popupbridgeexample://popupbridgev1/return#something=foo&other=bar")!
-        let _ = POPPopupBridge.open(url: mockURL)
+        let mockURL = URL(string: "sdk.ios.popup-bridge://popupbridgev1/return#something=foo&other=bar")!
+        mockWebAuthenticationSession.cannedResponseURL = mockURL
+
         let expectedResult = "window.popupBridge.onComplete(null, {\"path\":\"\\/return\",\"hash\":\"something=foo&other=bar\",\"queryItems\":{}});"
         let result = pub.constructJavaScriptCompletionResult(returnURL: mockURL)
 
         XCTAssertEqual(result, expectedResult)
     }
 
-    func testOpenURL_whenReturnURLHasNoURLFragment_passesPayloadWithNilHashToWebView() {
+    func testConstructJavaScriptCompletionResult_whenReturnURLHasNoURLFragment_passesPayloadWithNilHashToWebView() {
         let stubMessageBody: [String: String] = ["url": "http://example.com/?hello=world"]
         let stubMessageName = scriptMessageHandlerName
         let stubMessage = MockScriptMessage()
         stubMessage.body = stubMessageBody
         stubMessage.name = stubMessageName
 
-        let delegate = MockPopupBridgeDelegate()
-        let pub = POPPopupBridge(webView: WKWebView(), urlScheme: returnURL, delegate: delegate)
+        let pub = POPPopupBridge(webView: WKWebView(), webAuthenticationSession: mockWebAuthenticationSession)
 
         pub.userContentController(WKUserContentController(), didReceive: stubMessage)
 
-        let mockURL = URL(string: "com.braintreepayments.popupbridgeexample://popupbridgev1/return")!
-        let _ = POPPopupBridge.open(url: mockURL)
+        let mockURL = URL(string: "sdk.ios.popup-bridge://popupbridgev1/return")!
+        mockWebAuthenticationSession.cannedResponseURL = mockURL
+
         let expectedResult = "window.popupBridge.onComplete(null, {\"path\":\"\\/return\",\"queryItems\":{}});"
         let result = pub.constructJavaScriptCompletionResult(returnURL: mockURL)
 
         XCTAssertEqual(result, expectedResult)
     }
 
-    func testOpenURL_whenReturnURLDoesNotMatchScheme_returnsFalseAndDoesNotCallOnComplete() {
+    func testConstructJavaScriptCompletionResult_whenReturnURLDoesNotMatchScheme_returnsFalseAndDoesNotCallOnComplete() {
         let stubMessageBody: [String: String] = ["url": "http://example.com/?hello=world"]
         let stubMessageName = scriptMessageHandlerName
         let stubMessage = MockScriptMessage()
         stubMessage.body = stubMessageBody
         stubMessage.name = stubMessageName
 
-        let delegate = MockPopupBridgeDelegate()
-        let pub = POPPopupBridge(webView: WKWebView(), urlScheme: returnURL, delegate: delegate)
-
+        let pub = POPPopupBridge(webView: WKWebView())
         pub.userContentController(WKUserContentController(), didReceive: stubMessage)
 
-        let result = POPPopupBridge.open(url: URL(string: "not.the.right.scheme://popupbridgev1/return?something=foo")!)
-        XCTAssertFalse(result)
+        let result = pub.constructJavaScriptCompletionResult(returnURL: URL(string: "not.the.right.scheme://popupbridgev1/return?something=foo")!)
+        XCTAssertNil(result)
     }
 
-    func testOpenURL_whenReturnURLDoesNotMatchHost_returnsFalseAndDoesNotCallOnComplete() {
+    func testConstructJavaScriptCompletionResult_whenReturnURLDoesNotMatchHost_returnsFalseAndDoesNotCallOnComplete() {
         let stubMessageBody: [String: String] = ["url": "http://example.com/?hello=world"]
         let stubMessageName = scriptMessageHandlerName
         let stubMessage = MockScriptMessage()
         stubMessage.body = stubMessageBody
         stubMessage.name = stubMessageName
 
-        let delegate = MockPopupBridgeDelegate()
-        let pub = POPPopupBridge(webView: WKWebView(), urlScheme: returnURL, delegate: delegate)
-
+        let pub = POPPopupBridge(webView: WKWebView())
         pub.userContentController(WKUserContentController(), didReceive: stubMessage)
 
-        let result = POPPopupBridge.open(url: URL(string: "com.braintreepayments.popupbridgeexample://notcorrect/return?something=foo")!)
-        XCTAssertFalse(result)
-    }
-
-    func testDelegate_whenWebViewCallsPopupBridgeSendMessage_receivesMessage() {
-        let delegate = MockPopupBridgeDelegate()
-        let webView = WKWebView()
-        webView.navigationDelegate = self
-
-        let pub = POPPopupBridge(webView: webView, urlScheme: returnURL, delegate: delegate)
-        let expectation = expectation(description: "Called JS")
-
-        webViewReadyBlock = {
-            webView.evaluateJavaScript("window.popupBridge.sendMessage('myMessageName', JSON.stringify({foo: 'bar'}));") {_, _ in
-                delegate.popupBridge(pub, receivedMessage: "myMessageName", data: "{\"foo\":\"bar\"}")
-                XCTAssertEqual(delegate.receivedMessage, "myMessageName")
-                XCTAssertEqual(delegate.recievedData, "{\"foo\":\"bar\"}")
-                expectation.fulfill()
-            }
-        }()
-
-        webView.loadHTMLString("<html></html>", baseURL: nil)
-
-        waitForExpectations(timeout: 10)
+        let result = pub.constructJavaScriptCompletionResult(returnURL: URL(string: "sdk.ios.popup-bridge://popupbridgev2/return?something=foo")!)
+        XCTAssertNil(result)
     }
 
     // Consider adding tests for query parameter parsing - multiple values, special characters, encoded, etc.