+++ CHANGELOG.md
@@ -7,7 +7,11 @@
   * Require iOS 14+, Xcode 14.3+, and Swift 5.8+
   * Remove deprecated `POPPopupBridge.open(url:sourceApplication:)` & `POPPopupBridge.open(url:options:)` methods
   * Remove `POPPopupBridge.set(returnURLScheme:)`
-  * Add required `urlScheme` parameter to `POPPopupBridge.init()`
+  * Remove `open(url:)`
+  * Replace `SFSafariViewController` with `ASWebAuthenticationSession`
+  * No longer need to register a URL type to use Popup Bridge
+  * Removed `POPPopupBridgeDelegate`
+  * The initializer for `POPPopupBridge` now only requires a `WKWebView` `POPPopupBridge(webView:)`
 
 ## 1.2.0 (2021-01-22)